move lifecycle usages  to own modules.
