+++ lib/beacon/loader/page_module_loader.ex
@@ -1,5 +1,6 @@
 defmodule Beacon.Loader.PageModuleLoader do
   require Logger
+  alias Beacon.Lifecycle
   alias Beacon.Loader
   alias Beacon.Pages.Page
   alias Beacon.Pages.PageEvent
@@ -41,7 +42,7 @@ defmodule Beacon.Loader.PageModuleLoader do
 
   defp store_page(%Page{} = page, page_module, component_module) do
     %{id: page_id, layout_id: layout_id, site: site, path: path} = page
-    template = Beacon.Lifecycle.load_template(page)
+    template = Lifecycle.Template.load_template(page)
     Beacon.Router.add_page(site, path, {page_id, layout_id, page.format, template, page_module, component_module})
   end
