+++ lib/beacon_web/live/page_live.ex
@@ -3,6 +3,7 @@ defmodule BeaconWeb.PageLive do
   use Phoenix.HTML
   require Logger
   import Phoenix.Component
+  alias Beacon.Lifecycle
 
   def mount(:not_mounted_at_router, _params, socket) do
     {:ok, socket}
@@ -24,7 +25,7 @@ defmodule BeaconWeb.PageLive do
 
   def render(assigns) do
     {{site, path}, {_page_id, _layout_id, format, template, _page_module, _component_module}} = lookup_route!(assigns.__site__, assigns.__live_path__)
-    Beacon.Lifecycle.render_template(site: site, path: path, format: format, template: template, assigns: assigns, env: __ENV__)
+    Lifecycle.Template.render_template(site: site, path: path, format: format, template: template, assigns: assigns, env: __ENV__)
   end
 
   defp lookup_route!(site, path) do
